Move SleeperProperty to correct package
