+++ java/clients/src/main/java/sleeper/clients/admin/properties/PropertiesDiff.java
@@ -18,8 +18,8 @@ package sleeper.clients.admin.properties;
 
 import sleeper.clients.util.console.ConsoleOutput;
 import sleeper.configuration.properties.SleeperProperties;
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.SleeperPropertyIndex;
-import sleeper.configuration.properties.instance.SleeperProperty;
 
 import java.util.ArrayList;
 import java.util.Collections;
