+++ java/clients/src/main/java/sleeper/clients/admin/properties/UpdatePropertiesWithTextEditor.java
@@ -19,9 +19,9 @@ import sleeper.clients.util.ClientUtils;
 import sleeper.clients.util.CommandRunner;
 import sleeper.configuration.properties.PropertyGroup;
 import sleeper.configuration.properties.SleeperProperties;
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.format.SleeperPropertiesPrettyPrinter;
 import sleeper.configuration.properties.instance.InstanceProperties;
-import sleeper.configuration.properties.instance.SleeperProperty;
 import sleeper.configuration.properties.table.TableProperties;
 
 import java.io.BufferedWriter;
