+++ java/clients/src/main/java/sleeper/clients/util/ComparePropertiesTemplates.java
@@ -18,9 +18,9 @@ package sleeper.clients.util;
 import sleeper.clients.admin.properties.PropertiesDiff;
 import sleeper.clients.util.console.ConsoleOutput;
 import sleeper.configuration.properties.SleeperProperties;
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.format.GeneratePropertiesTemplates;
 import sleeper.configuration.properties.instance.InstanceProperties;
-import sleeper.configuration.properties.instance.SleeperProperty;
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
