+++ java/configuration/src/main/java/sleeper/configuration/properties/PropertyGroup.java
@@ -16,8 +16,6 @@
 
 package sleeper.configuration.properties;
 
-import sleeper.configuration.properties.instance.SleeperProperty;
-
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
