+++ java/configuration/src/main/java/sleeper/configuration/properties/SleeperPropertiesValidationReporter.java
@@ -15,8 +15,6 @@
  */
 package sleeper.configuration.properties;
 
-import sleeper.configuration.properties.instance.SleeperProperty;
-
 import java.util.LinkedHashMap;
 import java.util.Map;
