+++ java/configuration/src/main/java/sleeper/configuration/properties/format/SleeperPropertiesPrettyPrinter.java
@@ -21,9 +21,9 @@ import org.apache.commons.lang.WordUtils;
 import sleeper.configuration.properties.PropertiesUtils;
 import sleeper.configuration.properties.PropertyGroup;
 import sleeper.configuration.properties.SleeperProperties;
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.instance.InstanceProperty;
 import sleeper.configuration.properties.instance.InstancePropertyGroup;
-import sleeper.configuration.properties.instance.SleeperProperty;
 import sleeper.configuration.properties.table.TableProperty;
 import sleeper.configuration.properties.table.TablePropertyGroup;
