+++ java/configuration/src/main/java/sleeper/configuration/properties/table/TablePropertyDefaultValue.java
@@ -15,9 +15,9 @@
  */
 package sleeper.configuration.properties.table;
 
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.instance.InstanceProperty;
-import sleeper.configuration.properties.instance.SleeperProperty;
 
 @FunctionalInterface
 public interface TablePropertyDefaultValue {
