+++ java/configuration/src/main/java/sleeper/configuration/properties/table/TablePropertyImpl.java
@@ -19,8 +19,8 @@ package sleeper.configuration.properties.table;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 import sleeper.configuration.properties.PropertyGroup;
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.instance.InstanceProperties;
-import sleeper.configuration.properties.instance.SleeperProperty;
 
 import java.util.Objects;
 import java.util.Optional;
