+++ java/configuration/src/test/java/sleeper/configuration/properties/table/DummyTableProperty.java
@@ -16,8 +16,8 @@
 package sleeper.configuration.properties.table;
 
 import sleeper.configuration.properties.PropertyGroup;
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.instance.InstanceProperties;
-import sleeper.configuration.properties.instance.SleeperProperty;
 
 public class DummyTableProperty implements TableProperty {
