+++ java/configuration/src/test/java/sleeper/configuration/properties/table/TablePropertiesTest.java
@@ -18,8 +18,8 @@ package sleeper.configuration.properties.table;
 import org.junit.jupiter.api.Test;
 
 import sleeper.configuration.properties.DummySleeperProperty;
+import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.instance.InstanceProperties;
-import sleeper.configuration.properties.instance.SleeperProperty;
 
 import java.util.Iterator;
 import java.util.List;
